Adding build tags to disable runServerProxy on android.
