+++ flashlight.go
@@ -13,18 +13,14 @@ import (
 	"os/signal"
 	"runtime"
 	"runtime/pprof"
-	"strconv"
 	"strings"
 	"time"
 
 	"github.com/getlantern/flashlight/client"
 	"github.com/getlantern/flashlight/config"
 	"github.com/getlantern/flashlight/log"
-	"github.com/getlantern/flashlight/server"
 	"github.com/getlantern/flashlight/statreporter"
-	"github.com/getlantern/flashlight/statserver"
 	"github.com/getlantern/flashlight/util"
-	"github.com/getlantern/go-igdman/igdman"
 )
 
 const (
@@ -251,78 +247,12 @@ func runClientProxy(cfg *config.Config) {
 	}
 }
 
-// Runs the server-side proxy
-func runServerProxy(cfg *config.Config) {
-	useAllCores()
-
-	if cfg.Portmap > 0 {
-		log.Debugf("Attempting to map external port %d", cfg.Portmap)
-		err := mapPort(cfg)
-		if err != nil {
-			log.Errorf("Unable to map external port: %s", err)
-			os.Exit(PortmapFailure)
-		}
-		log.Debugf("Mapped external port %d", cfg.Portmap)
-	}
-
-	srv := &server.Server{
-		Addr:         cfg.Addr,
-		ReadTimeout:  0, // don't timeout
-		WriteTimeout: 0,
-		Host:         cfg.AdvertisedHost,
-		CertContext: &server.CertContext{
-			PKFile:         config.InConfigDir("proxypk.pem"),
-			ServerCertFile: config.InConfigDir("servercert.pem"),
-		},
-	}
-	if cfg.StatsAddr != "" {
-		// Serve stats
-		srv.StatServer = &statserver.Server{
-			Addr: cfg.StatsAddr,
-		}
-	}
-	err := srv.ListenAndServe()
-	if err != nil {
-		log.Fatalf("Unable to run server proxy: %s", err)
-	}
-}
-
-func mapPort(cfg *config.Config) error {
-	parts := strings.Split(cfg.Addr, ":")
-
-	internalPort, err := strconv.Atoi(parts[1])
-	if err != nil {
-		return fmt.Errorf("Unable to parse local port: ")
-	}
-
-	internalIP := parts[0]
-	if internalIP == "" {
-		internalIP, err = determineInternalIP()
-		if err != nil {
-			return fmt.Errorf("Unable to determine internal IP: %s", err)
-		}
-	}
-
-	igd, err := igdman.NewIGD()
-	if err != nil {
-		return fmt.Errorf("Unable to get IGD: %s", err)
-	}
-
-	igd.RemovePortMapping(igdman.TCP, cfg.Portmap)
-	err = igd.AddPortMapping(igdman.TCP, internalIP, internalPort, cfg.Portmap, 0)
-	if err != nil {
-		return fmt.Errorf("Unable to map port with igdman %d: %s", cfg.Portmap, err)
-	}
-
-	return nil
-}
-
 // determineInternalIP determines the internal IP to use for mapping ports. It
 // does this by dialing a website on the public Internet and then finding out
 // the LocalAddr for the corresponding connection. This gives us an interface
 // that we know has Internet access, which makes it suitable for port mapping.
 func determineInternalIP() (string, error) {
-	conn, err := net.DialTimeout("tcp", "s3.amazonaws.com:443", 20 * time.Second)
+	conn, err := net.DialTimeout("tcp", "s3.amazonaws.com:443", 20*time.Second)
 	if err != nil {
 		return "", fmt.Errorf("Unable to determine local IP: %s", err)
 	}
